requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
Django==5.2.4
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.0
dnspython==2.7.0
mongoengine==0.29.1
PyJWT==2.10.1
pymongo==4.13.2
python-decouple==3.8
redis==6.2.0
sqlparse==0.5.3
